discord/requirements.txt
grpcio
discord.py
typer
